move string utils to lib
