+++ generators/base/support/jhipster7-context.js
@@ -13,7 +13,7 @@ import {
 import { getDBCExtraOption } from '../../spring-data-relational/support/database-data.js';
 import { getJdbcUrl, getR2dbcUrl } from '../../spring-data-relational/support/database-url.js';
 import { fieldTypes } from '../../../jdl/jhipster/index.js';
-import { upperFirstCamelCase } from './string.js';
+import { upperFirstCamelCase } from '../../../lib/utils/string.js';
 
 const { BYTES, BYTE_BUFFER } = fieldTypes.RelationalOnlyDBTypes;
