+++ lib/utils/string.spec.ts
