+++ lib/utils/string.ts
